Move qb components into components directory
