+++ frontend/src/metabase/query_builder/components/AggregationWidget.jsx
