+++ frontend/src/metabase/query_builder/components/ExpandableString.jsx
