+++ frontend/src/metabase/query_builder/components/GuiQueryEditor.jsx
