+++ frontend/src/metabase/query_builder/components/LimitWidget.jsx
