+++ frontend/src/metabase/query_builder/components/NativeQueryEditor.jsx
