+++ frontend/src/metabase/query_builder/components/QueryHeader.jsx
