+++ frontend/src/metabase/query_builder/components/QueryModeButton.jsx
