+++ frontend/src/metabase/query_builder/components/QueryVisualization.jsx
