+++ frontend/src/metabase/query_builder/components/QueryVisualizationObjectDetailTable.jsx
