+++ frontend/src/metabase/query_builder/components/SearchBar.jsx
