+++ frontend/src/metabase/query_builder/components/TimeGroupingPopover.jsx
