+++ frontend/src/metabase/query_builder/components/VisualizationResult.jsx
