+++ frontend/src/metabase/query_builder/components/VisualizationSettings.jsx
