+++ frontend/src/metabase/query_builder/components/dataref/DetailPane.jsx
