+++ frontend/src/metabase/query_builder/components/dataref/MetricPane.jsx
