+++ frontend/src/metabase/query_builder/components/dataref/SegmentPane.jsx
