+++ frontend/src/metabase/query_builder/components/dataref/TablePane.jsx
