+++ frontend/src/metabase/query_builder/components/expressions/ExpressionEditorTextfield.css
