+++ frontend/src/metabase/query_builder/components/filters/pickers/NumberPicker.jsx
